pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smartpaste-ai"
version = "1.0.0"
description = "A context-aware AI clipboard assistant that enriches content and saves organized markdown files"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "SmartPaste Contributors" }
]
keywords = ["clipboard", "ai", "markdown", "automation", "productivity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
]

dependencies = [
    "pyperclip>=1.8.2",
    "requests>=2.31.0",
    "readability-lxml>=0.8.1",
    "langdetect>=1.0.9",
    "sentence-transformers>=2.2.2",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0.1",
    "click>=8.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
]
gui = [
    "pystray>=0.19.4",
    "tkinter-tooltip>=2.2.0",
    "customtkinter>=5.2.0",
    "pillow>=10.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[project.urls]
Homepage = "https://abdhajjar.github.io/SmartPaste"
Repository = "https://github.com/AbdHajjar/SmartPaste"
Issues = "https://github.com/AbdHajjar/SmartPaste/issues"
Documentation = "https://github.com/AbdHajjar/SmartPaste/blob/main/README.md"
"Download" = "https://github.com/AbdHajjar/SmartPaste/releases/latest"

[project.scripts]
smartpaste = "smartpaste.main:main"

[tool.hatch.build.targets.wheel]
packages = ["smartpaste"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]